Fix epoch data type
We store (and process) epoch as guint64 internally, but this
function was accepting guint32 instead, which caused issues with
on i686.

+++ modulemd/v2/include/modulemd-2.0/modulemd-rpm-map-entry.h
@@ -51,7 +51,7 @@ G_DECLARE_FINAL_TYPE (ModulemdRpmMapEntry,
  */
 ModulemdRpmMapEntry *
 modulemd_rpm_map_entry_new (const gchar *name,
-                            guint32 epoch,
+                            guint64 epoch,
                             const gchar *version,
                             const gchar *release,
                             const gchar *arch);
